requirements.txt
streamlit
requests
pillow
pymupdf
gtts
pylint  # Latest version, no need to pin unless issues persist
